checkAnswer>isCorrect
Changed checkAnswer to isCorrect and moved to QuizManager. Successfully called it in the VC. When any of the buttons are pressed,  it calls isCorrect, and passes in the tag from whichever button was pressed (via sender.tag) The result of this function dictates the 2 paths. Before I was thinking "Call the function. If the function is true, do something. If false, something else.  Correct way is if (function). Since function returns a Bool, this turns into if True>, else>. It's an important semantic shift that I need to wrap my head around since I've seen it before.